requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
pydantic==2.3.0
python-dotenv==1.0.0
faster-whisper==0.9.0
huggingface-hub>=0.19.0
librosa==0.10.1
numpy==1.25.2
pandas==2.1.0
scikit-learn==1.3.0
scipy==1.11.2
transformers==4.32.1
pymongo==4.5.0
motor==3.3.0
python-multipart==0.0.6
streamlit==1.26.0
plotly==5.16.1
soundfile==0.12.1
tqdm==4.66.1
nltk==3.8.1 